functions/requirements.txt
pandas
sqlalchemy
pymysql
kfp
google-cloud-secret-manager